# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/purchases": {
            "get": {
                "description": "Retrieve a list of purchases",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Purchase"
                ],
                "summary": "List Purchases",
                "parameters": [
                    {
                        "type": "string",
                        "name": "bought_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "created_at",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "product_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "salesperson_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.PurchaseList"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.Error"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new purchase",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Purchase"
                ],
                "summary": "Create Purchase",
                "parameters": [
                    {
                        "description": "Purchase data",
                        "name": "Purchase",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Purchase"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.PurchaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.Error"
                        }
                    }
                }
            }
        },
        "/purchases/{id}": {
            "get": {
                "description": "Retrieve a purchase by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Purchase"
                ],
                "summary": "Get Purchase",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Purchase ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.PurchaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.Error"
                        }
                    }
                }
            },
            "put": {
                "description": "Update purchase details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Purchase"
                ],
                "summary": "Update Purchase",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Purchase ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated purchase data",
                        "name": "PurchaseUpdate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.PurchaseUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.PurchaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a purchase by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Purchase"
                ],
                "summary": "Delete Purchase",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Purchase ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Message"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/entity.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/entity.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Error": {
            "type": "object",
            "properties": {
                "error": {}
            }
        },
...
        "entity.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "entity.Purchase": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "string"
                },
                "purchase_item": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.PurchaseItem"
                    }
                },
                "purchased_by": {
                    "type": "string"
                },
                "supplier_id": {
                    "type": "string"
                }
            }
        },
        "entity.PurchaseItem": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "string"
                },
                "purchase_price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "entity.PurchaseItemReq": {
            "type": "object",
            "properties": {
                "product_id": {
                    "type": "string"
                },
                "purchase_price": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                },
                "total_price": {
                    "type": "number"
                }
            }
        },
        "entity.PurchaseList": {
            "type": "object",
            "properties": {
                "purchases": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.PurchaseResponse"
                    }
                }
            }
        },
        "entity.PurchaseResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "string"
                },
                "purchase_item": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.PurchaseItemReq"
                    }
                },
                "purchased_by": {
                    "type": "string"
                },
                "supplier_id": {
                    "type": "string"
                },
                "total_cost": {
                    "type": "number"
                }
            }
        },
        "entity.PurchaseUpdate": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "payment_method": {
                    "type": "string"
                },
                "supplier_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your bearer token here",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}